# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Nsangusa API
  version: 1.0.0
  description: Static language-neutral contract for implemented Spring MVC routes.
    Spring Security supplies the explicitly marked login and logout operations.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
- url: https://api.example.invalid
  description: Placeholder; configure per environment
...
security:
- sessionCookie: []
- basicAuth: []
paths:
...
  "/api/v1/articles":
    get:
      tags:
      - Articles
      operationId: listLatestArticles
      summary: List the latest published articles using the legacy array response
      description: >
        Returns only PUBLISHED articles, ordered by publishedAt descending and id ascending. The limit is clamped
        to 1 through 100 rather than rejected. Use /api/v1/articles/discovery for lightweight, total-aware
        pagination.
      security: []
      parameters:
      - name: limit
        in: query
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        '200':
          description: At most 100 published full article views.
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                type: array
                maxItems: 100
                items:
                  $ref: '#/components/schemas/Article'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
  "/api/v1/articles/{slug}":
    get:
      tags:
      - Articles
      operationId: getPublishedArticle
      summary: Get a published article by slug
      security: []
      parameters:
      - "$ref": "#/components/parameters/Slug"
      responses:
        '200':
          description: Published article
          headers:
            Cache-Control:
              schema:
                type: string
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Article"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/UnexpectedError"
  "/api/v1/admin/articles":
    get:
      tags: [Articles]
      operationId: listEditorialArticles
      summary: Page through articles for editorial review
      x-required-roles: [EDITOR, ADMINISTRATOR]
      parameters:
      - "$ref": "#/components/parameters/WorkspacePage"
      - "$ref": "#/components/parameters/WorkspaceSize"
      - name: state
        in: query
        schema:
          "$ref": "#/components/schemas/ArticleState"
      responses:
        '200':
          description: Current article queue, never publicly cached
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ArticlePage"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/UnexpectedError"
    post:
      tags:
      - Articles
      operationId: createManualArticle
      summary: Create a sourced article for editorial review
      x-required-roles:
      - EDITOR
      - ADMINISTRATOR
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      - "$ref": "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ArticleCommandRequest"
      responses:
        '201':
          description: Article created
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/IdResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '409':
          "$ref": "#/components/responses/Conflict"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/UnexpectedError"
  "/api/v1/articles/{slug}/image":
    get:
      tags: [Media]
      operationId: getPublishedArticleImage
      summary: Serve a selected illustration only for a published article
      security: []
      parameters:
      - "$ref": "#/components/parameters/Slug"
      - "$ref": "#/components/parameters/ImageVariant"
      responses:
        '200':
          description: Optimized image bytes; no-store
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/svg+xml:
              schema:
                type: string
                format: binary
        '400':
          "$ref": "#/components/responses/BadRequest"
        '404':
          description: No published article or selected image
        default:
          "$ref": "#/components/responses/UnexpectedError"
...
  "/api/v1/admin/articles/{id}/revisions":
    get:
      tags: [Articles]
      operationId: listArticleRevisions
      summary: Inspect full immutable revision snapshots; unavailable older snapshots are explicitly marked
      x-required-roles: [EDITOR, ADMINISTRATOR]
      parameters:
      - "$ref": "#/components/parameters/Id"
      - "$ref": "#/components/parameters/WorkspacePage"
      - "$ref": "#/components/parameters/WorkspaceSize"
      responses:
        '200':
          description: Private paginated revision history
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/RevisionPage"
        default:
          "$ref": "#/components/responses/UnexpectedError"
  "/api/v1/admin/articles/{id}/revisions/compare":
    get:
      tags: [Articles]
      operationId: compareArticleRevisions
      summary: Compare two stored article revisions
      x-required-roles: [EDITOR, ADMINISTRATOR]
      parameters:
      - "$ref": "#/components/parameters/Id"
      - name: from
        in: query
        required: true
        schema: {type: integer, minimum: 1}
      - name: to
        in: query
        required: true
        schema: {type: integer, minimum: 1}
      responses:
        '200':
          description: Both snapshots and all changed top-level fields
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/RevisionComparison"
        default:
          "$ref": "#/components/responses/UnexpectedError"
  "/api/v1/admin/articles/{id}/corrections":
    post:
      tags: [Articles]
      operationId: startArticleCorrection
      summary: Withdraw a previously published article and begin a correction requiring renewed approval
      description: Preserves canonical identity and original publication time. Republishing the approved correction does not trigger another newsletter.
      x-required-roles: [EDITOR, ADMINISTRATOR]
      parameters:
      - "$ref": "#/components/parameters/Id"
      - "$ref": "#/components/parameters/CsrfToken"
      - "$ref": "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expectedVersion, note]
              properties:
                expectedVersion: {type: integer, format: int64, minimum: 0}
                note: {type: string, minLength: 1, maxLength: 2000}
      responses:
        '204':
          description: Article withdrawn and correction revision created
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/UnexpectedError"
  "/api/v1/admin/articles/{id}":
    get:
      tags:
      - Articles
      operationId: getAdminArticle
      summary: Get an article for editorial review
      x-required-roles:
      - EDITOR
      - ADMINISTRATOR
      parameters:
      - "$ref": "#/components/parameters/Id"
      responses:
        '200':
          description: Article
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Article"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/UnexpectedError"
    put:
      tags:
      - Articles
      operationId: editArticle
      summary: Replace editable article content and return it to review
      x-required-roles:
      - EDITOR
      - ADMINISTRATOR
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      - "$ref": "#/components/parameters/Id"
      - "$ref": "#/components/parameters/ExpectedVersion"
      - "$ref": "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ArticleCommandRequest"
      responses:
        '204':
          description: Operation completed
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '409':
          "$ref": "#/components/responses/Conflict"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/UnexpectedError"
  "/api/v1/admin/articles/{id}/approve":
    post:
      tags:
      - Articles
      operationId: approveArticle
      summary: Approve an article for publication workflow
      x-required-roles:
      - EDITOR
      - ADMINISTRATOR
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      - "$ref": "#/components/parameters/ReviewedVersion"
      - "$ref": "#/components/parameters/Id"
      - "$ref": "#/components/parameters/CsrfToken"
      responses:
        '202':
          description: Transition accepted
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '409':
          "$ref": "#/components/responses/Conflict"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/UnexpectedError"
  "/api/v1/admin/articles/{id}/publish":
    post:
      tags:
      - Articles
      operationId: publishArticle
      summary: Publish an approved article immediately
      x-required-roles:
      - EDITOR
      - ADMINISTRATOR
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      - "$ref": "#/components/parameters/ReviewedVersion"
      - "$ref": "#/components/parameters/Id"
      - "$ref": "#/components/parameters/CsrfToken"
      responses:
        '204':
          description: Transition applied
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '409':
          "$ref": "#/components/responses/Conflict"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/UnexpectedError"
  "/api/v1/admin/articles/{id}/reject":
    post:
      tags:
      - Articles
      operationId: rejectArticle
      summary: Reject an article awaiting review
      x-required-roles:
      - EDITOR
      - ADMINISTRATOR
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      - "$ref": "#/components/parameters/ReviewedVersion"
      - "$ref": "#/components/parameters/Id"
      - "$ref": "#/components/parameters/CsrfToken"
      responses:
        '204':
          description: Transition applied
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '409':
          "$ref": "#/components/responses/Conflict"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/UnexpectedError"
  "/api/v1/admin/articles/{id}/unpublish":
    post:
      tags:
      - Articles
      operationId: unpublishArticle
      summary: Remove a published article from public surfaces
      x-required-roles:
      - EDITOR
      - ADMINISTRATOR
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      - "$ref": "#/components/parameters/ReviewedVersion"
      - "$ref": "#/components/parameters/Id"
      - "$ref": "#/components/parameters/CsrfToken"
      responses:
        '204':
          description: Transition applied
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '409':
          "$ref": "#/components/responses/Conflict"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/UnexpectedError"
  "/api/v1/admin/articles/{id}/restore":
    post:
      tags:
      - Articles
      operationId: restoreArticle
      summary: Restore an unpublished article
      x-required-roles:
      - EDITOR
      - ADMINISTRATOR
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      - "$ref": "#/components/parameters/ReviewedVersion"
      - "$ref": "#/components/parameters/Id"
      - "$ref": "#/components/parameters/CsrfToken"
      responses:
        '204':
          description: Transition applied
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '409':
          "$ref": "#/components/responses/Conflict"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/UnexpectedError"
  "/api/v1/admin/articles/{id}/archive":
    post:
      tags:
      - Articles
      operationId: archiveArticle
      summary: Archive an unpublished or rejected article
      x-required-roles:
      - EDITOR
      - ADMINISTRATOR
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      - "$ref": "#/components/parameters/ReviewedVersion"
      - "$ref": "#/components/parameters/Id"
      - "$ref": "#/components/parameters/CsrfToken"
      responses:
        '204':
          description: Transition applied
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '409':
          "$ref": "#/components/responses/Conflict"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/UnexpectedError"
...
  "/api/v1/admin/articles/{articleId}/images/fallback":
    post:
      tags: [Media]
      operationId: requestArticleFallbackImage
      summary: Create an original neutral fallback illustration for explicit editorial review
      description: Does not invoke an AI provider or approve the image. A separate image approval is required before selection.
      x-required-roles: [EDITOR, ADMINISTRATOR]
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      - "$ref": "#/components/parameters/ArticleId"
      - "$ref": "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ImageFallbackRequest"
      responses:
        '202':
          description: Fallback image candidate created for review
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/EventResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          description: Article not found
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
        '409':
          "$ref": "#/components/responses/Conflict"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/UnexpectedError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: Redis-backed opaque Spring Session cookie
    basicAuth:
      type: http
      scheme: basic
  parameters:
...
    WorkspacePage:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    WorkspaceSize:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    Slug:
      name: slug
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 250
...
  responses:
    BadRequest:
      description: Invalid path, query, header, or request-body input
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
...
    Unauthorized:
      description: Authentication is required or failed
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
...
    TooManyRequests:
      description: Per-client request rate exceeded
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
      headers:
        Retry-After:
          schema:
            type: integer
          description: Retry delay in seconds
    ServiceUnavailable:
      description: Rate limiting could not reach its Redis dependency
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    UnexpectedError:
      description: Unexpected API failure
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
...